+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2026-03-30]
+### Fixed
+- Updated the `install-afc.sh` script to implement a `safe_copy` feature to prevent accidental overwriting of existing configuration files in 
+  certain circumstances.
+- Fixed an issue where using a custom unit name when installing would not properly update the unit name for buffer configurations.
+
 ## [2026-03-27]
 ### Added
 - Toolchanger: Restoring previous temperature when asynchronously loading standalone toolheads